Major structure reshuffling (#4)
