+++ src/raytrax/tracer/__init__.py
@@ -1,0 +1,1 @@
+"""Internal ODE integration layer: ray state types, solver, and JAX buffer types."""
